move commands to references controller
